Moved CompletedTest: Noticed that the model does not have a field for the test
