+++ profiles_api/completed_test/completed_test_api_view.py
@@ -1,0 +1,22 @@
+from rest_framework.views import APIView
+from rest_framework.response import Response
+from rest_framework import viewsets
+from rest_framework.authentication import TokenAuthentication
+from rest_framework.permissions import IsAuthenticated
+
+from profiles_api import permissions
+
+from profiles_api.completed_test.completed_test_serializer import CompletedTestSerializer
+from profiles_api.completed_test.completed_test_model import CompletedTest
+
+
+class CompletedTestViewSet(viewsets.ModelViewSet):
+    """Handles creating, reading and updating completed tests"""
+    authentication_classes = (TokenAuthentication,)
+    serializer_class = CompletedTestSerializer
+    queryset = CompletedTest.objects.all()
+    permission_classes = (permissions.UpdateOwnStatus, IsAuthenticated)
+
+    def perform_create(self, serializer):
+        """Sets the user profile to the logged in user"""
+        serializer.save(user_profile=self.request.user)
