+++ profiles_api/urls.py
@@ -9,6 +9,7 @@ from profiles_api.subtopic import subtopic_api_view
 from profiles_api.question import question_api_view
 from profiles_api.answer import answer_api_view
 from profiles_api.test import test_api_view
+from profiles_api.completed_test import completed_test_api_view
 
 
 router = DefaultRouter()
@@ -20,7 +21,7 @@ router.register('subtopic', subtopic_api_view.SubtopicViewSet)
 router.register('topic', topic_api_view.TopicViewSet)
 router.register('answer', answer_api_view.AnswerViewSet)
 router.register('test', test_api_view.TestViewSet)
-router.register('completedTest', views.CompletedTestViewSet)
+router.register('completedTest', completed_test_api_view.CompletedTestViewSet)
 router.register('theoryPage', views.TheoryPageViewSet)
