+++ profiles_api/views.py
@@ -127,18 +127,6 @@ class UserProfileFeedViewSet(viewsets.ModelViewSet):
         serializer.save(user_profile=self.request.user)
 
 
-class CompletedTestViewSet(viewsets.ModelViewSet):
-    """Handles creating, reading and updating completed tests"""
-    authentication_classes = (TokenAuthentication,)
-    serializer_class = serializers.CompletedTestSerializer
-    queryset = models.CompletedTest.objects.all()
-    permission_classes = (permissions.UpdateOwnStatus, IsAuthenticated)
-
-    def perform_create(self, serializer):
-        """Sets the user profile to the logged in user"""
-        serializer.save(user_profile=self.request.user)
-
-
 class TheoryPageViewSet(viewsets.ModelViewSet):
     """Handles creating, reading and updating theory pages"""
     authentication_classes = (TokenAuthentication,)
